// frontend/app/src/models/vikendica.ts
export default class Vikendica {
    id : number = 0;
    naziv: string = "";
    mesto: string = "";
    usluge: string = "";
    cenaLeto : number = 0;
    cenaZima : number = 0;
    gSirina: number = 0; 
    gDuzina: number = 0;
    slike: string[] = [];
    vlasnik : string = "";
    komentari : string[] = [];
    ocena : number[] = [];
    blokiranoDo : string = ""
}